增加时间段
